+++ llm/inference/__init__.py
@@ -1,4 +1,5 @@
 from llm.inference.base import InferenceEngine
 from llm.inference.multiproc import MultiprocessEngine
-from llm.inference.transformer import LogitsProcessorConfig, TransformersEngine
+from llm.inference.transformer import TransformersEngine
+from llm.inference.utils import LogitsProcessorConfig
 from llm.inference.vllm_client import VLLMClient
